Rename class name and test name
