+++ src/Setup/PluginManageCli.php
